requirements.txt
django
django-ninja
geojson
shapely
httpx
pipreqs